+++ src/org/lukasj/idea/torquescript/completion/TSLocalVariableCompletionContributor.kt
@@ -4,15 +4,15 @@ import com.intellij.codeInsight.completion.CompletionParameters
 import com.intellij.codeInsight.completion.CompletionProvider
 import com.intellij.codeInsight.completion.CompletionResultSet
 import com.intellij.codeInsight.lookup.LookupElementBuilder
-import com.intellij.psi.PsiElementResolveResult
 import com.intellij.psi.util.PsiTreeUtil
+import com.intellij.psi.util.elementType
 import com.intellij.psi.util.parentOfTypes
 import com.intellij.util.PlatformIcons
 import com.intellij.util.ProcessingContext
 import org.lukasj.idea.torquescript.psi.TSFile
 import org.lukasj.idea.torquescript.psi.TSTypes
-import org.lukasj.idea.torquescript.psi.TSVarExpression
 import org.lukasj.idea.torquescript.psi.impl.TSFunctionStatementElementImpl
+import org.lukasj.idea.torquescript.psi.impl.TSVarExpressionElementImpl
 
 
 class TSLocalVariableCompletionContributor : CompletionProvider<CompletionParameters>() {
@@ -23,7 +23,9 @@ class TSLocalVariableCompletionContributor : CompletionProvider<CompletionParame
     ) {
         val functionParent = parameters.position.parentOfTypes(TSFunctionStatementElementImpl::class)
         if (functionParent != null) {
-            PsiTreeUtil.findChildrenOfType(functionParent, TSVarExpression::class.java)
+            PsiTreeUtil.findChildrenOfType(functionParent, TSVarExpressionElementImpl::class.java)
+                .plus(functionParent.getParameters()
+                    .filter { it.elementType == TSTypes.LOCALVAR || it.elementType == TSTypes.THISVAR })
         } else {
             // Assume file-scoped
             (parameters.originalFile as TSFile).getVariables()