+++ src/org/lukasj/idea/torquescript/psi/impl/TSFunctionCallExpressionElementImpl.kt
@@ -10,6 +10,8 @@ import com.intellij.psi.util.siblings
 import com.intellij.refactoring.suggested.startOffset
 import org.lukasj.idea.torquescript.psi.*
 import org.lukasj.idea.torquescript.reference.TSFunctionCallReference
+import org.lukasj.idea.torquescript.reference.TSFunctionReference
+import org.lukasj.idea.torquescript.reference.TSObjectReference
 
 abstract class TSFunctionCallExpressionElementImpl(node: ASTNode) : ASTWrapperPsiElement(node),
     TSNamedElement {
@@ -17,21 +19,24 @@ abstract class TSFunctionCallExpressionElementImpl(node: ASTNode) : ASTWrapperPs
     abstract fun getExpression(): TSExpression
 
     override fun getNameIdentifier(): PsiElement? =
-        when {
-            getFunctionType() == TSFunctionType.METHOD -> {
+        when (functionType) {
+            TSFunctionType.METHOD -> {
                 if (getExpression() is TSIdentExpression) {
                     getExpression()
                 } else {
                     null
                 }
             }
-            getFunctionType() == TSFunctionType.GLOBAL -> {
+            TSFunctionType.GLOBAL -> {
                 firstChild
             }
-            getFunctionType() == TSFunctionType.GLOBAL_NS -> {
-                getExpression()
+            TSFunctionType.GLOBAL_NS -> {
+                if (getExpression() is TSIdentExpression) {
+                    getExpression()
+                } else {
+                    null
+                }
             }
-            else -> null
         }
 
     override fun getName(): String? =
@@ -46,33 +51,58 @@ abstract class TSFunctionCallExpressionElementImpl(node: ASTNode) : ASTWrapperPs
         return this
     }
 
-    fun getFunctionType(): TSFunctionType {
-        val expression = getExpression()
+    val functionType: TSFunctionType
+        get() {
+            val expression = getExpression()
 
-        if (expression.elementType == TSTypes.IDENT_EXPRESSION) {
-            return if (expression.node.findChildByType(TSTypes.COLON_COLON) != null) {
-                TSFunctionType.GLOBAL_NS
-            } else {
-                TSFunctionType.GLOBAL
+            if (expression.elementType == TSTypes.IDENT_EXPRESSION) {
+                return if (expression.node.findChildByType(TSTypes.COLON_COLON) != null) {
+                    TSFunctionType.GLOBAL_NS
+                } else {
+                    TSFunctionType.GLOBAL
+                }
             }
-        }
 
-        // It's an expression, we have no way of knowing what it might evaluate to..
-        return TSFunctionType.METHOD
-    }
+            // It's an expression, we have no way of knowing what it might evaluate to..
+            return TSFunctionType.METHOD
+        }
 
     override fun getReference(): PsiReference? {
         val identifier = nameIdentifier ?: return null
-        return when(getFunctionType()) {
+        return when (functionType) {
             TSFunctionType.GLOBAL -> {
-                TSFunctionCallReference(this, TextRange(0, identifier.textLength))
+                TSFunctionReference(this, TextRange(0, identifier.textLength))
             }
             TSFunctionType.GLOBAL_NS -> {
-                TSFunctionCallReference(this, TextRange(0, identifier.textLength))
+                TSFunctionReference(
+                    this,
+                    TextRange(
+                        identifier.lastChild.startOffsetInParent,
+                        identifier.lastChild.startOffsetInParent + identifier.lastChild.textLength
+                    )
+                )
             }
             TSFunctionType.METHOD -> {
-                TSFunctionCallReference(this, TextRange(0, identifier.textLength))
+                TSFunctionReference(
+                    this,
+                    TextRange(
+                        identifier.lastChild.startOffsetInParent,
+                        identifier.lastChild.startOffsetInParent + identifier.lastChild.textLength
+                    )
+                )
             }
+        }
+    }
+
+    override fun getReferences(): Array<PsiReference> {
+        val identifier = nameIdentifier ?: return arrayOf()
+        return if (identifier.firstChild != identifier.lastChild) {
+            arrayOf(
+                reference!!,
+                TSObjectReference(identifier.firstChild, TextRange(0, identifier.firstChild.textLength))
+            )
+        } else {
+            arrayOf(reference!!)
         }
     }
 }